# Dump: add performance timing