--- a/cmd/dumper.go
+++ b/cmd/dumper.go
@@ -42,7 +42,7 @@ func (d *dumper) bo() *backoff.ExponentialBackOff {
 }
 
 func (d *dumper) Dump(name string, v interface{}) {
-	w := tabwriter.NewWriter(os.Stdout, 0, 0, 1, ' ', 0)
+	w := tabwriter.NewWriter(os.Stdout, 0, 16, 2, ' ', 0)
 
 	var isHeating bool
 	if fd, ok := v.(api.FeatureDescriber); ok {
@@ -65,7 +65,7 @@ func (d *dumper) Dump(name string, v interface{}) {
 		if err != nil {
 			fmt.Fprintf(w, "Power:\t%v\n", err)
 		} else {
-			fmt.Fprintf(w, "Power:\t%.0fW\t(%v)\n", power, duration.Round(time.Millisecond))
+			fmt.Fprintf(w, "Power:\t%.0fW\t%v\n", power, duration.Round(time.Millisecond))
 		}
 	}
 
@@ -77,7 +77,7 @@ func (d *dumper) Dump(name string, v interface{}) {
 		if err != nil {
 			fmt.Fprintf(w, "Energy:\t%v\n", err)
 		} else {
-			fmt.Fprintf(w, "Energy:\t%.1fkWh\t(%v)\n", energy, duration.Round(time.Millisecond))
+			fmt.Fprintf(w, "Energy:\t%.1fkWh\t%v\n", energy, duration.Round(time.Millisecond))
 		}
 	}
 
@@ -89,7 +89,7 @@ func (d *dumper) Dump(name string, v interface{}) {
 		if err != nil {
 			fmt.Fprintf(w, "Current L1..L3:\t%v\n", err)
 		} else {
-			fmt.Fprintf(w, "Current L1..L3:\t%.3gA %.3gA %.3gA\t(%v)\n", i1, i2, i3, duration.Round(time.Millisecond))
+			fmt.Fprintf(w, "Current L1..L3:\t%.3gA %.3gA %.3gA\t%v\n", i1, i2, i3, duration.Round(time.Millisecond))
 		}
 	}
 
@@ -101,7 +101,7 @@ func (d *dumper) Dump(name string, v interface{}) {
 		if err != nil {
 			fmt.Fprintf(w, "Voltage L1..L3:\t%v\n", err)
 		} else {
-			fmt.Fprintf(w, "Voltage L1..L3:\t%.3gV %.3gV %.3gV\t(%v)\n", u1, u2, u3, duration.Round(time.Millisecond))
+			fmt.Fprintf(w, "Voltage L1..L3:\t%.3gV %.3gV %.3gV\t%v\n", u1, u2, u3, duration.Round(time.Millisecond))
 		}
 	}
 
@@ -113,7 +113,7 @@ func (d *dumper) Dump(name string, v interface{}) {
 		if err != nil {
 			fmt.Fprintf(w, "Power L1..L3:\t%v\n", err)
 		} else {
-			fmt.Fprintf(w, "Power L1..L3:\t%.0fW %.0fW %.0fW\t(%v)\n", p1, p2, p3, duration.Round(time.Millisecond))
+			fmt.Fprintf(w, "Power L1..L3:\t%.0fW %.0fW %.0fW\t%v\n", p1, p2, p3, duration.Round(time.Millisecond))
 		}
 	}
 
@@ -139,13 +139,13 @@ func (d *dumper) Dump(name string, v interface{}) {
 			if err != nil {
 				fmt.Fprintf(w, "Temp:\t%v\n", err)
 			} else {
-				fmt.Fprintf(w, "Temp:\t%.0f°C\t(%v)\n", soc, duration.Round(time.Millisecond))
+				fmt.Fprintf(w, "Temp:\t%.0f°C\t%v\n", soc, duration.Round(time.Millisecond))
 			}
 		} else {
 			if err != nil {
 				fmt.Fprintf(w, "Soc:\t%v\n", err)
 			} else {
-				fmt.Fprintf(w, "Soc:\t%.0f%%\t(%v)\n", soc, duration.Round(time.Millisecond))
+				fmt.Fprintf(w, "Soc:\t%.0f%%\t%v\n", soc, duration.Round(time.Millisecond))
 			}
 		}
 	}
@@ -168,7 +168,7 @@ func (d *dumper) Dump(name string, v interface{}) {
 		if err != nil {
 			fmt.Fprintf(w, "Charge status:\t%v\n", err)
 		} else {
-			fmt.Fprintf(w, "Charge status:\t%v\t(%v)\n", status, duration.Round(time.Millisecond))
+			fmt.Fprintf(w, "Charge status:\t%v\t%v\n", status, duration.Round(time.Millisecond))
 		}
 	}
 
@@ -180,7 +180,7 @@ func (d *dumper) Dump(name string, v interface{}) {
 		if err != nil {
 			fmt.Fprintf(w, "Status reason:\t%v\n", err)
 		} else {
-			fmt.Fprintf(w, "Status reason:\t%v\t(%v)\n", status, duration.Round(time.Millisecond))
+			fmt.Fprintf(w, "Status reason:\t%v\t%v\n", status, duration.Round(time.Millisecond))
 		}
 	}
 
@@ -197,7 +197,7 @@ func (d *dumper) Dump(name string, v interface{}) {
 		if err != nil {
 			fmt.Fprintf(w, "Enabled:\t%v\n", err)
 		} else {
-			fmt.Fprintf(w, "Enabled:\t%t\t(%v)\n", enabled, duration.Round(time.Millisecond))
+			fmt.Fprintf(w, "Enabled:\t%t\t%v\n", enabled, duration.Round(time.Millisecond))
 		}
 	}
 
@@ -209,7 +209,7 @@ func (d *dumper) Dump(name string, v interface{}) {
 		if err != nil {
 			fmt.Fprintf(w, "Charged:\t%v\n", err)
 		} else {
-			fmt.Fprintf(w, "Charged:\t%.1fkWh\t(%v)\n", energy, duration.Round(time.Millisecond))
+			fmt.Fprintf(w, "Charged:\t%.1fkWh\t%v\n", energy, duration.Round(time.Millisecond))
 		}
 	}
 
@@ -221,7 +221,7 @@ func (d *dumper) Dump(name string, v interface{}) {
 		if err != nil {
 			fmt.Fprintf(w, "Duration:\t%v\n", err)
 		} else {
-			fmt.Fprintf(w, "Duration:\t%v\t(%v)\n", chargeDuration.Truncate(time.Second), queryDuration.Round(time.Millisecond))
+			fmt.Fprintf(w, "Duration:\t%v\t%v\n", chargeDuration.Truncate(time.Second), queryDuration.Round(time.Millisecond))
 		}
 	}
 
@@ -233,7 +233,7 @@ func (d *dumper) Dump(name string, v interface{}) {
 		if err != nil {
 			fmt.Fprintf(w, "Mix/Max Current:\t%v\n", err)
 		} else {
-			fmt.Fprintf(w, "Mix/Max Current:\t%.1f/%.1fA\t(%v)\n", min, max, duration.Round(time.Millisecond))
+			fmt.Fprintf(w, "Mix/Max Current:\t%.1f/%.1fA\t%v\n", min, max, duration.Round(time.Millisecond))
 		}
 	}
 
@@ -247,7 +247,7 @@ func (d *dumper) Dump(name string, v interface{}) {
 		if err != nil {
 			fmt.Fprintf(w, "Range:\t%v\n", err)
 		} else {
-			fmt.Fprintf(w, "Range:\t%vkm\t(%v)\n", rng, duration.Round(time.Millisecond))
+			fmt.Fprintf(w, "Range:\t%vkm\t%v\n", rng, duration.Round(time.Millisecond))
 		}
 	}
 
@@ -259,7 +259,7 @@ func (d *dumper) Dump(name string, v interface{}) {
 		if err != nil {
 			fmt.Fprintf(w, "Odometer:\t%v\n", err)
 		} else {
-			fmt.Fprintf(w, "Odometer:\t%.0fkm\t(%v)\n", odo, duration.Round(time.Millisecond))
+			fmt.Fprintf(w, "Odometer:\t%.0fkm\t%v\n", odo, duration.Round(time.Millisecond))
 		}
 	}
 
@@ -271,7 +271,7 @@ func (d *dumper) Dump(name string, v interface{}) {
 		if err != nil {
 			fmt.Fprintf(w, "Finish time:\t%v\n", err)
 		} else {
-			fmt.Fprintf(w, "Finish time:\t%v\t(%v)\n", ft.Truncate(time.Minute).In(time.Local), duration.Round(time.Millisecond))
+			fmt.Fprintf(w, "Finish time:\t%v\t%v\n", ft.Truncate(time.Minute).In(time.Local), duration.Round(time.Millisecond))
 		}
 	}
 
@@ -283,7 +283,7 @@ func (d *dumper) Dump(name string, v interface{}) {
 		if err != nil {
 			fmt.Fprintf(w, "Climater:\t%v\n", err)
 		} else {
-			fmt.Fprintf(w, "Climate active:\t%v\t(%v)\n", active, duration.Round(time.Millisecond))
+			fmt.Fprintf(w, "Climate active:\t%v\t%v\n", active, duration.Round(time.Millisecond))
 		}
 	}
 
@@ -295,7 +295,7 @@ func (d *dumper) Dump(name string, v interface{}) {
 		if err != nil {
 			fmt.Fprintf(w, "Position:\t%v\n", err)
 		} else {
-			fmt.Fprintf(w, "Position:\t%v,%v\t(%v)\n", lat, lon, duration.Round(time.Millisecond))
+			fmt.Fprintf(w, "Position:\t%v,%v\t%v\n", lat, lon, duration.Round(time.Millisecond))
 		}
 	}
 
@@ -308,13 +308,13 @@ func (d *dumper) Dump(name string, v interface{}) {
 			if err != nil {
 				fmt.Fprintf(w, "Max Temp:\t%v\n", err)
 			} else {
-				fmt.Fprintf(w, "Max Temp:\t%d°C\t(%v)\n", limitSoc, duration.Round(time.Millisecond))
+				fmt.Fprintf(w, "Max Temp:\t%d°C\t%v\n", limitSoc, duration.Round(time.Millisecond))
 			}
 		} else {
 			if err != nil {
 				fmt.Fprintf(w, "Limit Soc:\t%v\n", err)
 			} else {
-				fmt.Fprintf(w, "Limit Soc:\t%d%%\t(%v)\n", limitSoc, duration.Round(time.Millisecond))
+				fmt.Fprintf(w, "Limit Soc:\t%d%%\t%v\n", limitSoc, duration.Round(time.Millisecond))
 			}
 		}
 	}
@@ -338,7 +338,7 @@ func (d *dumper) Dump(name string, v interface{}) {
 		if err != nil {
 			fmt.Fprintf(w, "Max Current:\t%v\n", err)
 		} else {
-			fmt.Fprintf(w, "Max Current:\t%.1fA\t(%v)\n", f, duration.Round(time.Millisecond))
+			fmt.Fprintf(w, "Max Current:\t%.1fA\t%v\n", f, duration.Round(time.Millisecond))
 		}
 	}
 
@@ -350,7 +350,7 @@ func (d *dumper) Dump(name string, v interface{}) {
 		if err != nil {
 			fmt.Fprintf(w, "Phases:\t%v\n", err)
 		} else {
-			fmt.Fprintf(w, "Phases:\t%d\t(%v)\n", f, duration.Round(time.Millisecond))
+			fmt.Fprintf(w, "Phases:\t%d\t%v\n", f, duration.Round(time.Millisecond))
 		}
 	}
 
@@ -367,7 +367,7 @@ func (d *dumper) Dump(name string, v interface{}) {
 			if id == "" {
 				id = "<none>"
 			}
-			fmt.Fprintf(w, "Identifier:\t%s\t(%v)\n", id, duration.Round(time.Millisecond))
+			fmt.Fprintf(w, "Identifier:\t%s\t%v\n", id, duration.Round(time.Millisecond))
 		}
 	}
 
@@ -379,11 +379,10 @@ func (d *dumper) Dump(name string, v interface{}) {
 		}
 	}
 
-	w.Flush()
-
-	// Output total query time
 	totalDuration := time.Since(totalStart)
-	fmt.Printf("\nTotal query time: %v\n", totalDuration.Round(time.Millisecond))
+	fmt.Fprintf(w, "\nTotal time:\t\t%v\n", totalDuration.Round(time.Millisecond))
+
+	w.Flush()
 }
 
 func (d *dumper) DumpDiagnosis(v interface{}) {