Adds performance timing for all API queries to device dump output.

```bash
$ evcc meter --name mixed

Power:           2340W              125ms
Energy:          1234.5kWh
Current L1..L3:  10.2A 9.8A 10.5A   2ms
Voltage L1..L3:  230V 231V 229V     2ms
Power L1..L3:    2346W 2264W 2405W  1ms

Total time:                         130ms
```

## Summary by Sourcery

Instrument the device dump output to measure and display execution timings for all API queries, including per-query and total durations

Enhancements:
- Introduce formatDuration and measureTime helpers to capture and format API call durations
- Replace individual Dump print calls with measureTime for all supported interfaces to include per-query timings
- Add an overall timer to compute and display the total dump duration
- Adjust tabwriter padding and static field formatting for aligned timing columns