@andig 2025-10-04
Bitte tabulare Ausgabe für Lesbarkeit