@andig 2025-10-05
Die Totaltime ist falsch und ich würde auch vorschlagen die wegzulassen da sie ohnehin nichts bedeutet?