@andig 2025-10-05
> Finde ich wichtig.

Aber wofür? Die zugriffe gibts sonst nirgendwo.