@premultiply 2025-10-05
Um zu Analysezwecken sehen zu können wie lange es insgesamt gedauert hat bis alle Daten da sind.