@andig 2025-10-05
Falls es nicht unbedingt ms sein müssen: bei mir sagt die aufrufende Shell, wie lange das Programm gebraucht hat.