@premultiply 2025-10-05
Die wenigstens Anwender dürften so eine Shell eingerichtet haben und selbst dann sagt sie auch nur wie lange das Programm gelaufen ist (inkl. Initialisierung & Co.) und nicht wie lange (nur) die Abfragen insgesamt gedauert haben.