@premultiply 2025-10-05
@sourcery-ai summary